Move no data rate limit from StreamExecution to ProgressReporter
